Trait rasp_leds::patterns::common::Pattern

pub trait Pattern {
    fn tick_rate(&self) -> u64;

    fn tick_cycle(&self) -> Option<u64>;

    fn tick(
        &mut self, 
        tick: u64, 
        leds: &mut [[u8; 4]]
    ) -> Result<TickAction, String>;

    fn init(&self, _runner: &RunnerInner) -> Result<(), String> { ... }

    fn start_tick(
        &mut self, 
        raw_tick: u64, 
        leds: &mut [[u8; 4]]
    ) -> Result<TickAction, String> { ... }
}

Required methods

fn tick_rate(&self) -> u64

Define the number of ms between ticks

fn tick_cycle(&self) -> Option<u64>

Defines tick_cycle which is passed into start_tick

fn tick(
&mut self,
tick: u64,
leds: &mut [[u8; 4]]
) -> Result<TickAction, String>

This is where the tick behavior needs to be implemented, modify the mutable leds array and then return a tickresult to decide how the runner handles it.

Provided methods

fn init(&self, _runner: &RunnerInner) -> Result<(), String>

Called when pattern is loaded into runner, result is returned over the API

fn start_tick(
&mut self,
raw_tick: u64,
leds: &mut [[u8; 4]]
) -> Result<TickAction, String>

Runs a single tick, doesnt neeed to be overwritten